Added `set` and `get`
Closes https://github.com/PistonDevelopers/current/issues/8
